% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\docType{methods}
\name{l2e.normal.sd}
\alias{l2e.normal.sd}
\alias{gat}
\alias{tf.dp}
\alias{segment.summarize}
\alias{prof2invals}
\alias{seg2invals}
\alias{weighted.ecf}
\alias{ecf.global.max}
\alias{expected.peak.heights}
\alias{quantum.sd}
\alias{irises.pluspurple}
\alias{timeseries.iod}
\title{Functions for processing raw profiles}
\format{
An object of class \code{character} of length 8.
}
\usage{
l2e.normal.sd(xs)

gat(x, iod)

tf.dp(y, lam)

segment.summarize(inprof, penalty, trans, seg, loc, se)

prof2invals(
  inprof,
  penalty,
  annotations,
  chrom.colname,
  bin.start.colname,
  bin.end.colname
)

seg2invals(seg_mean, seg_length, iod, annotations)

weighted.ecf(y, sds, s)

ecf.global.max(y, sds, smin = 1, smax = 8)

expected.peak.heights(cn, ratio.se, scale, svals)

quantum.sd(x, mu)

irises.pluspurple

timeseries.iod(v)
}
\description{
Functions for processing raw profiles
}
\keyword{internal}
